refactor: consolidate workflow conversation variable response type
